🔄 chore(CHANGELOG.md): update format and content of changelog The changelog has been updated to follow the Keep a Changelog format and adhere to Semantic Versioning. The previous release section has been removed as it did not follow the correct format. The changelog now includes a note that it was generated automatically by github_changelog_generator.
🔥 chore(README.md): remove generated content and update Table of Contents
The generated content and link to PDK documentation have been removed as they are not relevant to this project. The Table of Contents has been updated to reflect the current state of the README file.

📝 docs(README.md): update module description and setup instructions
The module now allows for the management of valid SSL certificates for the Puppet console via Lets Encrypt. The README has been updated to reflect this change. Additionally, the setup instructions have been updated to include a requirement to disable the default `http_redirect` and a warning that the module will cause a catalog compilation failure if it is not disabled.

📝 docs(README.md): update Usage section and remove deprecated Reference section
The Usage section now includes basic instructions for using the module, including adding it to the Puppetfile and classifying the puppet server. The deprecated Reference section has been removed, and users are encouraged to use Puppet Strings comments to generate a REFERENCE.md file.

🔗 chore(metadata.json): add source URL
The source URL has been added to the metadata.json file to provide a reference to the source code repository.